UI 설정 버그 수정
UI 설정 버그 수정

+++ Assets/01.Scripts/UI/ShopPanel.cs
@@ -9,7 +9,7 @@ public class ShopPanel : MonoBehaviour
 
     [SerializeField] private GameObject[] _lockedImages;
 
-    public float _bestScore;
+    public int _bestScore;
 
     public SpriteRenderer _rend;
     public Sprite[] Skins;
